Disallow new HIDL interfaces.
This saves a list of all HIDL interfaces and errors if new interfaces
are added. The Android T policy is that new AOSP interfaces must be
added in AIDL (ideally with full conversions, but optionally as
extension interfaces). This error is meant to avoid anyone accidentally
wasting time creating a new HIDL interface that cannot be accepted.

Now that is past the deadline for new Android S interfaces, adding this
in.

Ignore-AOSP-First: don't want to expose list of new HALs publicly,
  since I would want to check w/ all interface OWNERS first

Fixes: 182941786
Test: m nothing
Change-Id: Ieadef0dea7d30ac5085f8a087b540755f2717a69