Rework unit tests for lazy HIDL services.
With this commit, we introduce a new service that are used for testing
the active services callback.
This fixes a potential race in the tests and also introduces a better
way to check if the callback has effectively been run within the
service.
We pass a file descriptor to the service that is then used to notify the
test that the callback has successfully run and the service is about to
exit.
Previous tests only relied on the observation from the unit test that
the service is not running anymore.

Bug: 191781736
Test: atest hidl_lazy_test
Ignore-AOSP-First: already in AOSP, automerger issue workaround
Change-Id: Ifa0761f0f16862afc6b154abd151886d53ab1489